Setup db connection and one example api call
